Moving schema to types folder
