+++ src/tools/listPulseMetricDefinitions/listAllPulseMetricDefinitions.test.ts
@@ -1,6 +1,6 @@
 import { CallToolResult } from '@modelcontextprotocol/sdk/types.js';
 
-import type { PulseMetricDefinition } from '../../sdks/tableau/apis/pulseApi.js';
+import type { PulseMetricDefinition } from '../../sdks/tableau/types/pulse.js';
 import { server } from '../../server.js';
 import { listAllPulseMetricDefinitionsTool } from './listAllPulseMetricDefinitions.js';
