+++ src/tools/listPulseMetrics/listPulseMetricsFromMetricIds.test.ts
@@ -1,6 +1,6 @@
 import { CallToolResult } from '@modelcontextprotocol/sdk/types.js';
 
-import type { PulseMetric } from '../../sdks/tableau/apis/pulseApi.js';
+import type { PulseMetric } from '../../sdks/tableau/types/pulse.js';
 import { server } from '../../server.js';
 import { listPulseMetricsFromMetricIdsTool } from './listPulseMetricsFromMetricIds.js';
